+++ src/sage/algebras/all.py
@@ -30,6 +30,7 @@ from sage.algebras.quantum_groups.all import *
 from sage.algebras.lie_conformal_algebras.all import *
 
 from sage.algebras.flag_algebras import *
+from sage.algebras.combinatorial_theory import *
 
 # Algebra base classes
 from sage.algebras.free_algebra import FreeAlgebra
